requirements.txt
anndata==0.7.6
annoy==1.17.0
apipkg==1.5
apsw==3.9.2.post1
attrs==20.2.0
bbknn==1.3.12
bcrypt==3.2.0
cached-property==1.5.2
certifi==2020.6.20
cffi==1.14.3
cgatcore==0.6.7
cryptography==3.2.1
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
drmaa==0.7.9
et-xmlfile==1.0.1
execnet==1.7.1
fbpca==1.0
geosketch==1.1
get-version==2.1
gevent==20.9.0
greenlet==0.4.17
h5py==2.10.0
harmonypy==0.0.5
imageio==2.9.0
iniconfig==1.1.1
intervaltree==2.1.0
jdcal==1.4.1
joblib==0.17.0
kiwisolver==1.3.1
legacy-api-wrap==1.2
leidenalg==0.8.2
llvmlite==0.34.0
louvain==0.7.0
matplotlib==3.3.2
natsort==7.0.1
networkx==2.5
numba==0.51.2
numexpr==2.7.1
numpy==1.19.4
openpyxl==3.0.5
packaging==20.4
pandas==1.1.4
paramiko==2.7.2
patsy==0.5.1
pep8==1.7.1
Pillow==8.0.1
pluggy==0.13.1
py==1.9.0
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
pytest==6.1.2
pytest-cache==1.0
pytest-pep8==1.0.6
python-dateutil==2.8.1
python-igraph==0.8.3
pytz==2020.4
PyWavelets==1.1.1
PyYAML==5.3.1
ruffus==2.8.4
scanorama==1.6
scanpy==1.6.0
scikit-image==0.17.2
scikit-learn==0.23.2
scikit-misc==0.1.3
scipy==1.5.3
scrublet==0.2.1
seaborn==0.11.0
setuptools-scm==4.1.2
sinfo==0.3.1
six==1.15.0
sortedcontainers==2.2.2
SQLAlchemy==1.3.20
statsmodels==0.12.1
stdlib-list==0.7.0
tables==3.6.1
texttable==1.6.3
threadpoolctl==2.1.0
#tifffile==2020.10.1
toml==0.10.2
tqdm==4.51.0
umap-learn==0.4.6
zope.event==4.5.0
zope.interface==5.1.2
